requirements.txt
# Core Dependencies
python-dotenv==1.0.0
ccxt==4.2.25
asyncio==3.4.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Technical Analysis
pandas==2.2.0
numpy==1.26.3
ta==0.11.0

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
httpx==0.26.0

# Frontend (PyScript dependencies handled by browser)
aiofiles==23.2.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-json-logger==2.0.7

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0